+++ components.d.ts
@@ -15,7 +15,9 @@ declare module 'vue' {
     DashDropwdown: typeof import('./src/components/dropdowns/DashDropwdown.vue')['default']
     DashDropwDown: typeof import('./src/components/dropdowns/DashDropwDown.vue')['default']
     DashFormGenerator: typeof import('./src/components/forms/DashFormGenerator.vue')['default']
+    DashGrid: typeof import('./src/components/grids/DashGrid.vue')['default']
     DashInput: typeof import('./src/components/inputs/DashInput.vue')['default']
+    DashPanel: typeof import('./src/components/grids/DashPanel.vue')['default']
     DashSelect: typeof import('./src/components/selects/DashSelect.vue')['default']
     DashSwitch: typeof import('./src/components/checkboxes/DashSwitch.vue')['default']
     DashTagsInput: typeof import('./src/components/inputs/DashTagsInput.vue')['default']
